dispatcher: silently ignore empty files
There is already a way to hide/shadow scripts in "/usr/lib/NetworkManager/dispatcher.d":
by putting a file of the same name in "/etc/NetworkManager/dispatcher.d".

There is also the special case that if the file symlinks to "/dev/null", the
file is silently ignored. This is the proper way to hide a script.

I think we should also take a plain empty file as user indication to hide a script.
This way, one can simply hide a file with

  # touch /etc/NetworkManager/dispatcher.d/10-ifcfg-rh-routes.sh

It's an alternative to symlinking to /dev/null.