clients/tests: support replacing UUIDs in command text
We want to issue a nmcli command with the connection's UUID in the command
line. As the command line is stored to disk and compared, we cannot use UUIDs
that are randomly generated. Instead, we need to support replacing the text
in the command line.

(cherry picked from commit 7d048f1a5bb1af6aa8f05ad4a40e84b4da647731)