libnm: expose nm_context_busy_watcher_integrate_source() as internal API for reuse
The mechanics how NMClient handles the GMainContext and the
context-busy-watcher apply really to every GObject that uses
GDBusConnection and register to signals.

At least, as long as the API provides no shutdown/stop method,
because that means shutdown/stop happens when unreferencing the
instance, at which point pending operations get cancelled (but
they cannot complete right away due to the nature of
g_dbus_connection_call()). If you had a shutdown/stop API, then all
pending operations could keep the instance alive, and the instance
sticks around (and keeps the GMainContext busy) until shutdown is
completed. But in existing API that does not intend the user to
explicitly shutdown, that's not a feasible (backward compatible)
addition. But the context-busy-watcher object is.