libnm: move nm_client_get_capabilities() to separate linker version
nm_client_get_capabilities() was backported to 1.22.2. Add to to the
appropriate linker version.

Officially (and according to docs) nm_client_get_capabilities() still
appears first in libnm 1.24.0. However, as it got backported to 1.22.2,
it needs to be part of a different symbol version on 1.22. Instead
of adding the symbol twice (once for libnm_1_24_0 and libnm_1_22_2),
move it only to the libnm_1_22_2 symbol version, also on master.