license: relicense "shared/nm-std-aux/nm-dbus-compat.h" under LGPL-2.1+
"nm-dbus-compat.h" was GPL licensed. That is a problem, because we use it from
libnm (which is LGPL).

The history of this file in NetworkManager source tree:

  $ git shortlog -n -s e055bdbbc3e7bb3ff3bc73860c4479f15bbfe7b2 -- shared/nm-std-aux/nm-dbus-compat.h include/nm-dbus-compat.h shared/nm-dbus-compat.h
       5    Thomas Haller
       1    Dan Winship
       1    Lubomir Rintel

Note that commit dd0e19895532 ('include: add nm-dbus-compat.h')
introduced this file from dbus sources ([1]). Hence, originally
the file is (like all of dbus sources) dual-licensed under GPL-2.0+
and Academic Free License 2.1 (AFL-2.1). That makes it problematic to
change the license of this file to LGPL also because of the old history
of the file.

Instead, drop everything from the header except the bits that we
actually use. I claim the remainder is trivial and only contains
defines for documented D-Bus API. I don't think that the remainder
is copyrightable and hence get rid of the copy-right notice and the
GPL license.

[1] https://gitlab.freedesktop.org/dbus/dbus/blob/39ea37b587c9fa4c5e8d0aff08b043617b22f50a/dbus/dbus-shared.h